make skins properly themable
