+++ source/skins/simple/CheckboxSkin.js
@@ -1,33 +1,7 @@
 import React from 'react';
-import classnames from 'classnames';
-import { pickDOMProps } from '../../utils/props';
 import { themr } from 'react-css-themr';
 import { CHECKBOX } from './identifiers';
 import DefaultCheckboxTheme from '../../themes/simple/SimpleCheckbox.scss';
+import CheckboxSkin from "./raw/CheckboxSkin";
 
-export default themr(CHECKBOX, DefaultCheckboxTheme, { withRef: true })((props) => (
-  <div
-    className={classnames([
-      props.className,
-      props.theme.root,
-      props.disabled ? props.theme.disabled : null,
-      props.checked ? props.theme.checked : null,
-    ])}
-    onClick={(event) => {
-      if (!props.disabled && props.onChange) {
-        props.onChange(!props.checked, event);
-      }
-    }}
-  >
-    <input
-      {...pickDOMProps(props)}
-      className={props.theme.input}
-      type="checkbox"
-    />
-    <div className={classnames([
-      props.theme.check,
-      props.checked ? props.theme.checked : null,
-    ])} />
-    {props.label ? (<label className={props.theme.label}>{props.label}</label>) : null}
-  </div>
-));
+export default themr(CHECKBOX, DefaultCheckboxTheme)(CheckboxSkin);
