+++ source/skins/simple/FormFieldSkin.js
@@ -1,27 +1,7 @@
 import React from 'react';
-import classnames from 'classnames';
 import { themr } from 'react-css-themr';
 import { FORM_FIELD } from './identifiers';
 import DefaultFormFieldTheme from '../../themes/simple/SimpleFormField.scss';
+import FormFieldSkin from "./raw/FormFieldSkin";
 
-export default themr(FORM_FIELD, DefaultFormFieldTheme, { withRef: true })((props) => (
-  <div
-    className={classnames([
-      props.className,
-      props.theme.root,
-      props.disabled ? props.theme.disabled : null,
-      props.error ? props.theme.errored : null,
-    ])}
-  >
-    {props.error && <div className={props.theme.error}>{props.error}</div>}
-    {props.label && (
-      <label
-        className={props.theme.label}
-        onClick={props.component.focus ? props.component.focus : null}
-      >
-        {props.label}
-      </label>
-    )}
-    {props.input}
-  </div>
-));
+export default themr(FORM_FIELD, DefaultFormFieldTheme)(FormFieldSkin);
