+++ source/skins/simple/SwitchSkin.js
@@ -1,36 +1,7 @@
 import React from 'react';
-import classnames from 'classnames';
-import { pickDOMProps } from '../../utils/props';
 import { themr } from 'react-css-themr';
 import { SWITCH } from './identifiers';
 import DefaultSwitchTheme from '../../themes/simple/SimpleSwitch.scss';
+import SwitchSkin from "./raw/SwitchSkin";
 
-export default themr(SWITCH, DefaultSwitchTheme, { withRef: true })((props) => (
-  <div
-    className={classnames([
-      props.className,
-      props.theme.root,
-      props.disabled ? props.theme.disabled : null,
-      props.checked ? props.theme.checked : null,
-    ])}
-    onClick={(event) => {
-      if (!props.disabled && props.onChange) {
-        props.onChange(!props.checked, event);
-      }
-    }}
-  >
-    <input
-      {...pickDOMProps(props)}
-      className={props.theme.input}
-      readOnly
-      type="checkbox"
-    />
-    <div className={classnames([
-      props.theme.switch,
-      props.checked ? props.theme.checked : null,
-    ])}>
-      <span className={props.theme.thumb} />
-    </div>
-    {props.label ? (<label className={props.theme.label}>{props.label}</label>) : null}
-  </div>
-));
+export default themr(SWITCH, DefaultSwitchTheme)(SwitchSkin);
