+++ source/skins/simple/raw/ModalSkin.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import ReactModal from 'react-modal';
+import { themr } from 'react-css-themr';
+import { MODAL } from '../identifiers';
+
+export default themr(MODAL)((props) => (
+  <ReactModal
+    contentLabel={props.contentLabel}
+    isOpen={props.isOpen}
+    onRequestClose={props.onClose}
+    shouldCloseOnOverlayClick={props.triggerCloseOnOverlayClick}
+    className={props.theme.modal}
+    overlayClassName={props.theme.overlay}
+  >
+    {props.children}
+  </ReactModal>
+));
